requirements.txt
asgiref==3.8.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.3
django-extensions==3.2.3
django-import-export==3.3.7
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.2.0
psycopg2==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
tzdata==2024.1
xlrd==2.0.1
xlwt==1.3.0
